% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{addPlayers}
\alias{addPlayers}
\title{Add new players to the saved list}
\usage{
addPlayers(players, file = PLAYER_FILE)
}
\arguments{
\item{players}{vector with new player names}

\item{file}{Path to the RDS file with saved players}
}
\description{
Add new players to the saved list
}
