% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisionPetanque.R
\name{collisionPetanque}
\alias{collisionPetanque}
\title{Model collisions between petanque balls}
\usage{
collisionPetanque(data, throwBallID, throwBallV, a = 10)
}
\arguments{
\item{data}{Data.frame with position: 'x', 
r: 'radius' and 'm': mass of each ball}

\item{throwBallID}{ID of the thrown ball.}

\item{throwBallV}{Velocity of the thrown ball (m/s).}

\item{a}{Acceleration (currently fixed), decceleration of 10m/s^2 by default.}
}
\value{
List with data.frame with iterative positions of the balls (in case of successive collisions).
}
\description{
Model collisions between petanque balls
}
\examples{
# Example for the collisioPetanque function.
# 
# Author: Laure Cougnaud
###############################################################################

# size petanque field: 4 * 15m in length
mPetanque <- 0.680 # standards: 680 et 710 grammes
mTarget <- 0.34 # between 10 and 18 gramms
# from: https://www.petanqueshop.com/le-choix-d-une-boule-de-competition
radiusPetanque <- 0.075 # official competition: between 70,5 et 80 mm
radiusTarget <- 0.030 # 30 diameter +- 1mm

# position of thrown ball
x <- 7.9
v <- getSpeedFromDistance(d = x, theta = pi/4)

dataPetanque <- data.frame(
		ID = c(1:4),
		x = c(x, 2, 3, 8), # coordinates in m
		r = c(radiusPetanque, radiusPetanque, radiusPetanque, radiusTarget),
		m = c(mPetanque, mPetanque, mPetanque, mTarget)
)

collisionPetanque(data = dataPetanque, throwBallID = 1, throwBallV = v, a = 20)

}
\author{
Laure Cougnaud
}
