% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{computePWe}
\alias{computePWe}
\title{Special rating algorithm for unrated players, players with less than 8 games and player with all wins or all losses}
\usage{
computePWe(rankingPlayer, rankingPlayeri)
}
\arguments{
\item{players}{Player's name}

\item{rankingDB}{The current ranking database}
}
\description{
Special rating algorithm for unrated players, players with less than 8 games and player with all wins or all losses
}
