% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCollision.R
\name{detectCollision}
\alias{detectCollision}
\title{Detect collision}
\usage{
detectCollision(posDF, collisNo = 0)
}
\arguments{
\item{posDF}{Data.frame with balls position}

\item{collisNo}{Integer with collision number}
}
\value{
Updated \code{posDF}
}
\description{
Detect collision
}
\author{
Jason Waddell, small update Laure Cougnaud
}
