% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{getEffectiveNrGames}
\alias{getEffectiveNrGames}
\title{Get the number of effective games played by a player}
\usage{
getEffectiveNrGames(oldRanking, gamesPlayed)
}
\arguments{
\item{oldRanking}{Current ranking of a player}

\item{gamesPlayed}{Number of games played}
}
\description{
Get the number of effective games played by a player
}
