% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisionPetanque.R
\name{getKE}
\alias{getKE}
\title{Get Kinetic Energy}
\usage{
getKE(m, v)
}
\arguments{
\item{m}{Mass of the object in kg}

\item{v}{Object velocity, in m/s}
}
\value{
Kinetic Energy in Joules
}
\description{
Get Kinetic Energy
}
\author{
Laure Cougnaud
}
