% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{getPlayers}
\alias{getPlayers}
\title{Get a vector of all saved player names}
\usage{
getPlayers(file = PLAYER_FILE)
}
\arguments{
\item{file}{Path to the RDS file with saved players}
}
\description{
Get a vector of all saved player names
}
