% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{getRanking}
\alias{getRanking}
\title{Get the current ranking of a player or initialize a player's ranking}
\usage{
getRanking(player, rankingDB = NULL, file = RANKING_FILE)
}
\arguments{
\item{rankingDB}{The current ranking database}

\item{players}{Player's name}
}
\description{
Get the current ranking of a player or initialize a player's ranking
}
