% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisionPetanque.R
\name{getSpeedFromDistance}
\alias{getSpeedFromDistance}
\title{Get initial speed of the projectile from the distance}
\usage{
getSpeedFromDistance(d, theta = pi/4)
}
\arguments{
\item{d}{Distance in m}

\item{theta}{Launching angle, 45 by default}
}
\value{
Speed in m/s
}
\description{
From https://en.wikipedia.org/wiki/Projectile_motion, 
section: 'Maximum distance of projectile'
}
\author{
Laure Cougnaud
}
