% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collisionPetanque.R
\name{getVFromKE}
\alias{getVFromKE}
\title{Get velocity from Kinetic Energy and Mass}
\usage{
getVFromKE(KE, m)
}
\arguments{
\item{KE}{Kinetic Energy in Joules}

\item{m}{Mass of the object in kg}
}
\value{
Object velocity, in m/s
}
\description{
Get velocity from Kinetic Energy and Mass
}
\author{
Laure Cougnaud
}
