% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{objFunction}
\alias{objFunction}
\title{Objective function to minimize}
\usage{
objFunction(rank, effectiveGames, R_0_prime, ratingPlayer2, S_prime)
}
\arguments{
\item{rank}{Current rank}

\item{effectiveGames}{Number of effective games}

\item{R_0_prime}{Result of R_0_prime}

\item{ratingPlayer2}{The rating of player 2}

\item{S_prime}{Result of the S_prime parameter}

\item{rankingDB}{The current ranking database}
}
\description{
Objective function to minimize
}
