% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampleDistribution}
\alias{sampleDistribution}
\title{Sample possible distributions choices from the big list}
\usage{
sampleDistribution(n = 3)
}
\arguments{
\item{n}{Number of distributions to sample}
}
\description{
Sample possible distributions choices from the big list
}
