% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{specialRanking}
\alias{specialRanking}
\title{Special rating algorithm for unrated players, players with less than 8 games or all successive wins/losses}
\usage{
specialRanking(infoPlayer1, ratingPlayer2, score)
}
\arguments{
\item{infoPlayer1}{information of player 1}

\item{ratingPlayer2}{The rating of player 2}

\item{score}{Scored obtained by the player}
}
\description{
Special rating algorithm for unrated players, players with less than 8 games or all successive wins/losses
}
