% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{standardRanking}
\alias{standardRanking}
\title{Standard rating algorithm for players with more than 8 games}
\usage{
standardRanking(infoPlayer1, ratingPlayer2, score)
}
\arguments{
\item{infoPlayer1}{information of player 1}

\item{ratingPlayer2}{The rating of player 2}

\item{score}{Scored obtained by the player}
}
\description{
Standard rating algorithm for players with more than 8 games
}
