% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{updateRanking}
\alias{updateRanking}
\title{Update ranking based on the result of the game}
\usage{
updateRanking(players, winner, score, file = RANKING_FILE)
}
\arguments{
\item{players}{Player names (character vector of length 2)}

\item{winner}{Winning player's number 1 or 2}

\item{score}{Score}

\item{file}{Path to the RDS file with saved ranking}
}
\description{
Update ranking based on the result of the game
}
